% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_snake_case.R
\name{custom_snake_case}
\alias{custom_snake_case}
\title{Coerce strings into snake case formatting}
\usage{
custom_snake_case(x)
}
\arguments{
\item{x}{Vector of strings to coerce to snake case.}
}
\description{
This function takes vector of strings and turns each string to snake case format.
The function also replaces any \% symbols with "perc". Further modifiers can be added before
the function internally uses the snakecase::to_any_case(x, case = "snake").
}
\examples{
custom_snake_case(rownames(mtcars))
}
