% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_numeric_names.R
\name{guess_numeric_names}
\alias{guess_numeric_names}
\title{Identify which strings start with a number}
\usage{
guess_numeric_names(x)
}
\arguments{
\item{x}{Either data.table or data.frame with column names in snake format.}
}
\description{
This function allows to determine if columns in data.frame or data.table start with number.
It can only detect numbers that are either one or two digits and in snake case format.
For example, "20_cars" will be detected but "20 cars" or "200 cars" will not be detected.
Typical use case is to rename certain columns such as:
setnames(data, guess_numeric_names(data), paste0("name", guess_numeric_names(data)))
}
\examples{
colnames(mtcars) <- c("Name", "1_Works", "1_Works \%", "2 Fails", "2 Fails \%",
     "30_Works", "30_Works \%", "700_Fails", "700_Fails \%", "Longer", "Longer \%")

guess_numeric_names(mtcars)
}
\keyword{guess}
