% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rascut-package.R
\docType{package}
\name{rascut-package}
\alias{rascut}
\alias{rascut-package}
\title{rascut: Functions Used across Variety of Projects at the ASC Team in the DHSC}
\description{
The rascut acronym stands for R Adult Social Care (ASC) Utilities. These are collection of
     functions used across variety of projects at the ASC team in the Department of Health and Social Care. 
     The primary purpose is to provide single point of access to all functions and improve reproducibility 
     of these projects.
}
\author{
\strong{Maintainer}: Martin Cadek \email{martin.cadek@dhsc.co.uk}

}
\keyword{internal}
